Move drawer to dedicated component
